Fix the issue that crashes the socket
The issue was attempting to concatenate binaries with `<>` operator but
mistaking an integer (heartbeat) as one side of the binaries.

Surprisingly, the compiler didn't raise any error as `1 <> "a"` would
do. It just hung up for a few seconds and then crashed the connection.

This may have something to do with WebSockex package, since it has a
`try_callback` function in its loop that tries to catch all sorts of
errors.

[Issue log]

> *DBG* WS forcefully closed the connection because the server was taking too long close

> WS disconnected: %{attempt_number: 1, conn: %WebSockex.Conn{cacerts: nil, conn_mod: :ssl, extra_headers: [], host: "testnet.bitmex.com", insecure: true, path: "/realtime", port: 443, query: nil, socket: nil, socket_connect_timeout: 6000, socket_recv_timeout: 5000, transport: :ssl}, reason: {%ArgumentError{message: "argument error"}, [{:erlang, :bit_size, [217], []}, {WS, :handle_info, 2, [file: 'lib/bitmex/application.ex', line: 11]}, {WebSockex, :try_callback, 3, [file: 'lib/websockex.ex', line: 900]}, {WebSockex, :common_handle, 4, [file: 'lib/websockex.ex', line: 644]}, {:proc_lib, :init_p_do_apply, 3, [file: 'proc_lib.erl', line: 247]}]}}

> *DBG* WS received msg: {:ssl_closed, {:sslsocket, {:gen_tcp, #Port<0.6523>, :tls_connection, :undefined}, #PID<0.206.0>}}